Cleaning & Comments added
Appropriately modified the time taken display variable name.
Added comments on some of the methods, some incredibly simple methods were not commented due to simplicity.
Comments are not really needed on most of these methods but probably better practice to get into the habit of having them. Definitely worth honing the wording of these method summaries to be more concise and useful. As this is a personal project I have left in some comments that describe potential future features and other less useful information.