Full workday hours equaliser
Allows the user to specify certain tasks as having been done during a set workday, then creates a task that makes up the remainder of the hours of that day. Means tasks don't have to be adjusted to make up the full workday.
Works for any day of the week.

Also includes several small improvements to other areas of the application. Such as moving the OnPropertyChanged for TotalTimeTakenDisplay into the setter for loggedtasks, a check to ensure a timeitem exists before trying to delete it (won't try to delete index = -1), and the TotalTimeTakenDisplay is now blank on startup