Improvements
Hovering over the name or comments of a task shows a tooltip containing the full contents of the label, allowing the user to read overflowed text.
Also modified the CurrentTask duration tracker at the top left to show the actual full time taken of the task, not just the time taken since continuing it.